Tests for [WFCORE-3568] Allow CLI output scrolling
